[REFACTOR] : article detail viewHolder 분리 / #128
