+++ app/src/main/java/com/indipage/presentation/articledetail/ArticleDetailBottomAdapter.kt
@@ -3,14 +3,13 @@ package com.indipage.presentation.articledetail
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 import com.example.core_ui.view.ItemDiffCallback
-import com.indipage.data.dto.response.ResponseArticleDetailDto
 import com.indipage.databinding.ItemArticleDetailBottomBinding
 import com.indipage.domain.entity.ArticleDetail
+import com.indipage.presentation.articledetail.viewholder.ArticleDetailBottomViewHolder
 
 class ArticleDetailBottomAdapter() :
-    ListAdapter<ArticleDetail, ArticleDetailBottomAdapter.ArticleDetailBottomViewHolder>(
+    ListAdapter<ArticleDetail, ArticleDetailBottomViewHolder>(
         ArticleAllDiffCallback
     ) {
 
@@ -25,17 +24,6 @@ class ArticleDetailBottomAdapter() :
 
     override fun onBindViewHolder(holder: ArticleDetailBottomViewHolder, position: Int) {
         holder.bind(getItem(position))
-    }
-
-    class ArticleDetailBottomViewHolder(
-        private val binding: ItemArticleDetailBottomBinding,
-    ) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(data: ArticleDetail) {
-            with(binding) {
-                articleDetail = data
-                executePendingBindings()
-            }
-        }
     }
 
     companion object {
